More ci work (#49)
* Larger error tolerance for bar_layered_transparent

Layer order doesn't always match and doesn't seem well defined

* Cross compile arm wheels on macOS

* Resolve dependencies without default channel

* Only save CPython wheels for 3.8-3.10

* Change minimum Python version from 3.7 to 3.8

* Also trigger workflow when PR is opened

* Comment Rust tests and non-linux builds

* Try latest chrome version

* enable build mac wheels

* Fix tests path

* Remove unwanted wheels before uploading artifacts

* Comment mac wheel build

* Fix utc date parts return type

* Update wheel artifact logic

* Download and install wheels in test-vegafusion-jupyter job

* Change minimum Python version back to 3.7

* Install test dependencies from local repo

* fixes

* Remove stale notebooks

* Fix chart-editor example

* ci fix

* Move test dependencies to conda environment

* Downgrade python-chrome-binary to 96 to match CI

* Sleep more

* Run vegafusion-jupyter tests on mac and windows too

* Remove 3.8 f-string syntax

* use working-directory instead of cd for windows compat

* win -> mac

* Try mac with Python 3.7

* Add retrying to driver.get_url()

* Try moving voila process management to module level

* Support spec as Altair object, dict, or string

* Don't include _null in auto-generated aggregate field name

* Revert "Try moving voila process management to module level"

This reverts commit 21a4c7e56a289869de847e5563f6909409446fcc.

* Sleep after shutting down voila, use flaky to retry up to 3 times

* Comment out non-linux integration tests, restor packaging

* cargo fmt

+++ vegafusion-rt-datafusion/src/expression/compiler/builtin_functions/date_time/date_parts.rs
@@ -225,7 +225,7 @@ pub fn make_datepart_udf_utc(extract_fn: fn(&DateTime<Utc>) -> i64, name: &str) 
     };
     let part_fn = make_scalar_function(part_fn);
 
-    let return_type: ReturnTypeFunction = Arc::new(move |_| Ok(Arc::new(DataType::Int32)));
+    let return_type: ReturnTypeFunction = Arc::new(move |_| Ok(Arc::new(DataType::Int64)));
     ScalarUDF::new(
         name,
         &Signature::uniform(
